refactor: (pools): Move common pool modals from pools card to components
