+++ src/views/Pools/components/LockedPool/Buttons/AddCakeButton.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from '@pancakeswap/localization'
 import { usePool } from 'state/pools/hooks'
 import AddAmountModal from '../Modals/AddAmountModal'
 import { AddButtonProps } from '../types'
-import NotEnoughTokensModal from '../../PoolCard/Modals/NotEnoughTokensModal'
+import NotEnoughTokensModal from '../../Modals/NotEnoughTokensModal'
 
 const AddCakeButton: React.FC<React.PropsWithChildren<AddButtonProps>> = ({
   currentBalance,
