+++ src/views/Pools/components/PoolCard/CardActions/StakeActions.tsx
@@ -4,8 +4,8 @@ import { useTranslation } from '@pancakeswap/localization'
 import { getBalanceNumber } from 'utils/formatBalance'
 import { DeserializedPool } from 'state/types'
 import Balance from 'components/Balance'
-import NotEnoughTokensModal from '../Modals/NotEnoughTokensModal'
-import StakeModal from '../Modals/StakeModal'
+import NotEnoughTokensModal from '../../Modals/NotEnoughTokensModal'
+import StakeModal from '../../Modals/StakeModal'
 
 interface StakeActionsProps {
   pool: DeserializedPool
