+++ src/views/Pools/components/PoolsTable/ActionPanel/Harvest.tsx
@@ -9,7 +9,7 @@ import { BIG_ZERO } from 'utils/bigNumber'
 import { DeserializedPool } from 'state/types'
 
 import { ActionContainer, ActionTitles, ActionContent } from './styles'
-import CollectModal from '../../PoolCard/Modals/CollectModal'
+import CollectModal from '../../Modals/CollectModal'
 
 const HarvestAction: React.FunctionComponent<React.PropsWithChildren<DeserializedPool>> = ({
   sousId,
